+++ titiler/pgstac/main.py
@@ -5,7 +5,7 @@ from typing import Dict
 
 from titiler.core.dependencies import TileMatrixSetName, TMSParams
 from titiler.core.errors import DEFAULT_STATUS_CODES, add_exception_handlers
-from titiler.core.factory import TMSFactory
+from titiler.core.factory import MultiBaseTilerFactory, TMSFactory
 from titiler.core.middleware import (
     CacheControlMiddleware,
     LoggerMiddleware,
@@ -14,7 +14,9 @@ from titiler.core.middleware import (
 from titiler.core.resources.enums import OptionalHeader
 from titiler.mosaic.errors import MOSAIC_STATUS_CODES
 from titiler.pgstac.db import close_db_connection, connect_to_db
+from titiler.pgstac.dependencies import ItemPathParams
 from titiler.pgstac.factory import MosaicTilerFactory
+from titiler.pgstac.reader import PgSTACReader
 from titiler.pgstac.settings import ApiSettings
 from titiler.pgstac.version import __version__ as titiler_pgstac_version
 
@@ -68,8 +70,16 @@ else:
     optional_headers = []
 
 
-mosaic = MosaicTilerFactory(optional_headers=optional_headers)
-app.include_router(mosaic.router)
+mosaic = MosaicTilerFactory(optional_headers=optional_headers, router_prefix="/mosaic")
+app.include_router(mosaic.router, tags=["Mosaic"], prefix="/mosaic")
+
+stac = MultiBaseTilerFactory(
+    reader=PgSTACReader,
+    path_dependency=ItemPathParams,
+    optional_headers=optional_headers,
+    router_prefix="/stac",
+)
+app.include_router(stac.router, tags=["Items"], prefix="/stac")
 
 tms = TMSFactory(supported_tms=TileMatrixSetName, tms_dependency=TMSParams)
 app.include_router(tms.router, tags=["TileMatrixSets"])